backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
# Face shape detection
mediapipe==0.10.9
opencv-python-headless==4.8.1.78
numpy==1.26.2
pillow==10.1.0
# Chat Assistant API
openai>=1.0.0
